Convert code to Python 3
The "rb" argument used on line 18 was changed to "r" in order
to avoid an exception on line 45 related to the conditional
expression expect a bytes-type object rather than a string.
A generator expression was implementedin build_wordlist() for
efficiency and to strip the trailing "\n" from each path read
from the all.txt file.
- Refactored to comply with PEP8
- Reformatted indentation